'utils' module renamed to more specific 'version_compatibility';
