+++ preferences.py
@@ -4,7 +4,7 @@ import bpy
 from bpy.props import StringProperty
 
 from . import preset_handler
-from .utils import make_annotations, layout_split
+from .version_compatibility import make_annotations, layout_split
 
 
 @make_annotations
